backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
mysqlclient>=2.2
python-dotenv>=1.0
django-cors-headers>=4.0
django-environ>=0.11
PyJWT>=2.8
pymysql~=1.1.1
requests>=2.0
redis>=4.5.0
django-redis>=5.3.0
django-cacheops>=7.0.0
celery>=5.3.0
django-celery-beat>=2.5.0
gunicorn>=21.0.0
whitenoise>=6.5.0